+++ __tests__/10_deep_proxy_spec.js
@@ -1,4 +1,10 @@
-import { createDeepProxy, isDeepChanged } from '../src/deepProxy';
+import {
+  createDeepProxy,
+  isDeepChanged,
+  MODE_ASSUME_UNCHANGED_IF_UNAFFECTED,
+} from '../src/deepProxy';
+
+const MODE_DEFAULT = MODE_ASSUME_UNCHANGED_IF_UNAFFECTED;
 
 const noop = () => {
   // do nothing
@@ -10,9 +16,9 @@ describe('shallow object spec', () => {
     const a1 = new WeakMap();
     const p1 = createDeepProxy(s1, a1);
     noop(p1);
-    expect(isDeepChanged(s1, { a: 'a', b: 'b' }, a1)).toBe(false);
-    expect(isDeepChanged(s1, { a: 'a2', b: 'b' }, a1)).toBe(false);
-    expect(isDeepChanged(s1, { a: 'a', b: 'b2' }, a1)).toBe(false);
+    expect(isDeepChanged(s1, { a: 'a', b: 'b' }, a1, null, MODE_DEFAULT)).toBe(false);
+    expect(isDeepChanged(s1, { a: 'a2', b: 'b' }, a1, null, MODE_DEFAULT)).toBe(false);
+    expect(isDeepChanged(s1, { a: 'a', b: 'b2' }, a1, null, MODE_DEFAULT)).toBe(false);
   });
 
   it('one property access', () => {